requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-core==1.35.0
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
isodate==0.7.2
motor==3.7.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp==1.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20